# sistema-ventas/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Sistema de Ventas",
        "description": "Documentación de la API para el sistema de ventas",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de la API"
        }
    ],
    "paths": {
...
        "/api/sales": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Obtener todas las ventas",
                "description": "Devuelve una lista de todas las ventas registradas en el sistema",
                "operationId": "getAllSales",
                "responses": {
                    "200": {
                        "description": "Lista de ventas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "550e8400-e29b-41d4-a716-446655440000"
                                            },
                                            "sale_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-01-14"
                                            },
                                            "sale_details": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "product_id": {
                                                            "type": "string",
                                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                                        },
                                                        "quantity": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "unit_price": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 100.5
                                                        },
                                                        "subtotal": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 201
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Crear una nueva venta",
                "description": "Crea una nueva venta en el sistema",
                "operationId": "createSale",
                "requestBody": {
                    "description": "Datos de la venta a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sale_date",
                                    "sale_details"
                                ],
                                "properties": {
                                    "sale_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-14"
                                    },
                                    "sale_details": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "unit_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100.5
                                                },
                                                "subtotal": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 201
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Venta creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sale": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "550e8400-e29b-41d4-a716-446655440006"
                                                },
                                                "sale_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-01-14"
                                                },
                                                "sale_details": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "product_id": {
                                                                "type": "string",
                                                                "example": "550e8400-e29b-41d4-a716-446655440000"
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "unit_price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 100.5
                                                            },
                                                            "subtotal": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 201
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Venta creada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "El campo sale_date es obligatorio."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al crear la venta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}